pyproject.toml
[project]
name = "hastalavistamoney"
version = "1.7.0"
description = "Home Accounting Project"
authors = [
    {name = "Alexander Pavlov", email = "dev@pavlovteam.ru"}
]
requires-python = ">=3.12.7"
dependencies = [
    "dj-database-url>=3.0.0",
    "psycopg2-binary>=2.9.3",
    "whitenoise[brotli]>=6.0.0",
    "django-filter>=25.1",
    "PyYAML>=6.0",
    "django==5.2.7",
    "urllib3>=2.0.2",
    "django-csp>=4.0",
    "channels>=4.0.0",
    "Pillow>=11.0.0",
    "numpy>=2.0.0",
    "django-axes>=8.0.0",
    "Twisted[http2,tls]>=25.0.0",
    "python-dateutil>=2.8.2",
    "django-debug-toolbar>=5.0.0",
    "sentry-sdk>=2.22.0",
    "django-rosetta>=0.10.0",
    "crispy-bootstrap5>=2025.4",
    "django-crispy-forms>=2.1",
    "djangorestframework>=3.15.0",
    "django-stubs-ext>=5.1.1",
    "djangorestframework-simplejwt>=5.5.0",
    "openai>=1.84.0",
    "django-structlog>=9.1.1",
    "granian[reload]>=2.3.4",
    "django-extensions>=4.1",
    "djangorestframework-stubs>=3.16.1",
    "django-stubs>=5.2.1",
    "django-cors-headers>=4.7.0",
    "python-decouple>=3.8",
    "django-redis>=6.0.0",
    "django-compressor>=4.5.1",
]

[project.optional-dependencies]
dev = [
    "Faker>=37.0.0",
    "coverage>=7.3.2",
    "myst-parser>=4.0.0",
    "flower>=2.0.1",
    "pre-commit>=4.0.0",
    "yesqa>=1.5.0",
    "mypy>=1.5.1",
    "bump2version>=1.0.1",
    "icecream>=2.1.3",
    "werkzeug>=3.0.1",
    "django-extensions>=4.1",
    "ruff>=0.12.0",
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.29.0",
    "mkdocs-material>=9.5.39",
    "django-stubs[compatible-mypy]>=5.1.0",
    "types-python-dateutil>=2.9.0.20241003",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hasta_la_vista_money"]

[tool.ruff]
line-length = 80
exclude = [
    "**/migrations/*",
    "**/__init__",
    ".git",
    "__pycache__",
    "manage.py"
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
exclude = [
    "migrations",
    ".venv",
    ".git"
]

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = [
    "D100", "D101", "D102", "D104", "D105", "D106", "E203", "COM812", "Q000", "Q003", "PT009", "RUF001", "RUF002", "PLR0913",
]
external = [ "WPS" ]

[tool.ruff.lint.mccabe]
max-complexity = 6

[tool.ruff.lint.per-file-ignores]
"**/tests/*.py" = [
  "S101", # asserts
  "S105", # hardcoded passwords
  "S404", # subprocess calls are for tests
  "S603", # do not require `shell=True`
  "S607", # partial executable paths
]

[tool.mypy]
python_version = "3.13"
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
]
files = ["hasta_la_vista_money/"]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
warn_return_any = true
disallow_incomplete_defs = true
pretty = true
show_error_codes = true
exclude = [
  "migrations",
]

[[tool.mypy.overrides]]
module = "hasta_la_vista_money.*.migrations.*"
ignore_errors = true

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.pytest.ini_options]
norecursedirs = ["__pycache__"]
addopts = "--strict-markers"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
default_section = "FIRSTPARTY"
line_length = 79

[tool.bumpversion]
current_version = "1.7.0"
commit = false
tag = true

[[tool.bumpversion.files]]
filename = "hasta_la_vista_money/__init__.py"
search = "__version__ = '{current_version}'"
replace = "__version__ = '{new_version}'"

[tool.bumpversion."file:setup.cfg"]
search = "current_version = {current_version}"
replace = "current_version = {new_version}"

[tool.bumpversion."file:pyproject.toml"]
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[mypy.plugins.django-stubs]
django_settings_module = "config.django.base"

[dependency-groups]
dev = [
    "faker>=37.4.0",
]
